Refactor to integrate MongoDB models and operations; remove old models and operations files
